# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: One Finance UX Hub API
  version: 0.2.0
  description: >
    Event ingest, the stitch fold, and the read/write APIs the console calls.
    The browser uses only these HTTP + SSE endpoints; it never touches the bus.
    Worked example everywhere: REV-ACC / FOBO / R-1042 READY / R-2031 BLOCKED.
servers:
  - url: http://localhost:7070
    description: Local hub
...
paths:
...
  /api/stitch/instances:
    get:
      tags: [instances]
      summary: Board / list, with fold counts. Every filter is a query parameter.
      parameters:
        - { name: groupUnit, in: query, schema: { type: string } }
        - { name: cobDate, in: query, schema: { type: string, format: date } }
        - { name: region, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string, enum: [NOT_YET, READY, BLOCKED, CLEARED, DELAYED] } }
      responses:
        '200':
          description: Instances
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Instance' }
...
  /api/stitch/notifications:
    get:
      tags: [notifications]
      summary: Notification inbox (instance-linked)
      parameters: [ { name: limit, in: query, schema: { type: integer, default: 50 } } ]
      responses:
        '200':
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Notification' }
...
  /api/stitch/deadletters/{id}/replay:
    post:
      tags: [rtb]
      summary: Dual-control replay of a held dead letter
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      requestBody:
        content:
          application/json:
            schema: { type: object, properties: { by: { type: string } } }
      responses:
        '200': { description: Replayed }
...
  /api/stitch/kits:
    get:
      tags: [onboarding]
      summary: Registered product kits
      responses:
        '200':
          description: Kits
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Kit' } }
    post:
      tags: [onboarding]
      summary: Register a product kit as data (no new code)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/KitRequest' }
      responses:
        '201': { description: Kit registered }
        '400': { description: Unknown source/destination or duplicate kit }
...
components:
  schemas:
...
    Instance:
      type: object
      properties:
        instanceId: { type: string }
        kitId: { type: string }
        groupUnitId: { type: string }
        question: { type: string }
        sliceKey: { type: string }
        region: { type: string }
        cobDate: { type: string, format: date }
        status: { type: string }
        runId: { type: string, nullable: true }
        namedBlocker: { type: string, nullable: true }
        openEscalations: { type: integer }
        completed: { type: integer }
        failed: { type: integer }
        waiting: { type: integer }
        expected: { type: integer }
...
    Notification:
      type: object
      properties:
        id: { type: integer }
        createdAt: { type: string, format: date-time }
        severity: { type: string }
        transition: { type: string }
        outcomeKey: { type: string }
        outcomeName: { type: string }
        audience: { type: string }
        title: { type: string }
        message: { type: string }
        channels: { type: string }
        instanceId: { type: string, nullable: true }
        groupUnitId: { type: string, nullable: true }
...
    Source:
      type: object
      properties:
        sourceId: { type: string }
        displayName: { type: string }
        identifierType: { type: string }
        ingestMode: { type: string }
        topicOrUrl: { type: string, nullable: true }
        mapperId: { type: string, nullable: true }
    Destination:
      type: object
      properties:
        destId: { type: string }
        displayName: { type: string }
        actionType: { type: string }
        commandUrl: { type: string, nullable: true }
        stepOrder: { type: integer, nullable: true }
    Kit:
      type: object
      properties:
        kitId: { type: string }
        groupUnitId: { type: string }
        domain: { type: string }
        question: { type: string }
        renderer: { type: string }
        userActions: { type: string }
        ceesProduct: { type: string }
        slaCutoff: { type: string, nullable: true }
        status: { type: string }
        sources: { type: array, items: { $ref: '#/components/schemas/Source' } }
        destinations: { type: array, items: { $ref: '#/components/schemas/Destination' } }
    KitRequest:
      type: object
      required: [kitId, groupUnitId, question, renderer, userActions, ceesProduct]
      properties:
        kitId: { type: string }
        groupUnitId: { type: string }
        domain: { type: string }
        question: { type: string }
        renderer: { type: string, enum: [HELIX_RECON, ENGINE_REPORT, GRID_PACK, NOTIFY_MILESTONE, ANALYST_VIEW] }
        userActions: { type: string, example: SIGN_OFF,POST }
        ceesProduct: { type: string }
        slaCutoff: { type: string, nullable: true }
        sources:
          type: array
          items:
            type: object
            properties: { sourceId: { type: string }, required: { type: boolean } }
        destinations:
          type: array
          items:
            type: object
            properties: { destId: { type: string }, stepOrder: { type: integer } }
        embed:
          type: object
          nullable: true
          properties:
            url: { type: string }
            allowedOrigin: { type: string }
            chrome: { type: string }